Refactoring Configuracion
Implementación de gestión de clientes:
- Se añadió funcionalidad de paginación, búsqueda y ordenación en la tabla de clientes.
- Se implementó contexto global para clientes con `ClienteContext`.
- Se agregaron opciones de edición y eliminación de clientes.

+++ src/Configuracion/ListadoCliente.jsx
@@ -15,43 +15,46 @@ export default function ListadoCliente() {
     direction: "ascending",
   });
 
+  // Carga los datos al montar el componente o cambiar la paginación / búsqueda
   useEffect(() => {
     getDatos();
   }, [page, pageSize, consulta]);
 
-  const handlePageChange = (newPage) => {
-    setPage(newPage);
-  };
-
+  // Función para obtener datos
   const getDatos = async () => {
     console.log("carga " + page);
-    obtenerClientes(consulta, page, pageSize)
-      .then((response) => {
-        setClientes(response.content);
-        setTotalPages(response.totalPages);
-      })
-      .catch((error) => {
-        console.error("Error fetching items:", error);
-      });
+    try {
+      const response = await obtenerClientes(consulta, page, pageSize);
+      setClientes(response.content);
+      setTotalPages(response.totalPages);
+    } catch (error) {
+      console.error("Error fetching items:", error);
+    }
   };
 
+  // Función para manejar el cambio de la búsqueda
   const handConsultaChange = (e) => {
     setConsulta(e.target.value);
   };
 
+  // Función para eliminar un cliente
   const eliminar = async (id) => {
-    try {
-      const eliminacionExitosa = await eliminarCliente(id);
-      if (eliminacionExitosa) {
-        getDatos();
-      } else {
-        console.error("Error al eliminar el cliente");
+    const confirmacion = window.confirm("¿Estás seguro de eliminar este cliente?");
+    if (confirmacion) {
+      try {
+        const eliminacionExitosa = await eliminarCliente(id);
+        if (eliminacionExitosa) {
+          getDatos();
+        } else {
+          console.error("Error al eliminar el cliente");
+        }
+      } catch (error) {
+        console.error("Error al eliminar el cliente:", error);
       }
-    } catch (error) {
-      console.error("Error al eliminar el cliente:", error);
     }
   };
 
+  // Función para manejar la ordenación de columnas
   const handleSort = (key) => {
     let direction = "ascending";
     if (sortConfig.key === key && sortConfig.direction === "ascending") {
@@ -60,6 +63,7 @@ export default function ListadoCliente() {
     setSortConfig({ key, direction });
   };
 
+  // Función para ordenar los datos
   const sortedData = () => {
     const sorted = [...clientes];
     if (sortConfig.key !== null) {
@@ -75,15 +79,16 @@ export default function ListadoCliente() {
     }
     return sorted;
   };
-  ///////////////////////////////////////Hasta aca para el orden de las tablas///////////////////////////////////////////////////
+
+  // Función para manejar el cambio de página
+  const handlePageChange = (newPage) => {
+    setPage(newPage);
+  };
 
   return (
     <div className="container">
-      <div>
-        <h1> Gestión de Clientes </h1>
-        <hr></hr>
-      </div>
-
+      <h1> Gestión de Clientes </h1>
+      <hr />
       <div className="row d-md-flex justify-content-md-end">
         <div className="col-5">
           <input
@@ -94,88 +99,67 @@ export default function ListadoCliente() {
             aria-label="Search"
             value={consulta}
             onChange={handConsultaChange}
-          ></input>
+          />
         </div>
         <div className="col-1">
-          <button
-            onClick={() => getDatos()}
-            className="btn btn-outline-success"
-            type="submit"
-          >
+          <button onClick={getDatos} className="btn btn-outline-success">
             Buscar
           </button>
         </div>
       </div>
-      <hr></hr>
+      <hr />
       <table className="table table-striped table-hover align-middle">
         <thead className="table-dark">
           <tr>
             <th scope="col" onClick={() => handleSort("id")}>
               #
               {sortConfig.key === "id" && (
-                <span>
-                  {sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}
-                </span>
+                <span>{sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}</span>
               )}
             </th>
             <th scope="col" onClick={() => handleSort("razonSocial")}>
               Apellido y Nombre
               {sortConfig.key === "razonSocial" && (
-                <span>
-                  {sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}
-                </span>
+                <span>{sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}</span>
               )}
             </th>
             <th scope="col" onClick={() => handleSort("celular")}>
               Cel
               {sortConfig.key === "celular" && (
-                <span>
-                  {sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}
-                </span>
+                <span>{sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}</span>
               )}
             </th>
             <th scope="col" onClick={() => handleSort("mail")}>
               Mail
               {sortConfig.key === "mail" && (
-                <span>
-                  {sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}
-                </span>
+                <span>{sortConfig.direction === "ascending" ? " 🔽" : " 🔼"}</span>
               )}
             </th>
             <th scope="col">Acciones</th>
           </tr>
         </thead>
         <tbody>
-          {
-            //iteramos empleados
-            sortedData().map((cliente, indice) => (
-              <tr key={indice}>
-                <th scope="row">{cliente.id}</th>
-                <td>{cliente.razonSocial}</td>
-                <td>{cliente.celular}</td>
-                <td>{cliente.mail}</td>
-
-                <td className="text-center">
-                  <div>
-                    <Link
-                      to={`/cliente/${cliente.id}`}
-                      className="btn btn-link btn-sm me-3"
-                    >
-                      Editar
-                    </Link>
-
-                    <button
-                      onClick={() => eliminar(cliente.id)}
-                      className="btn btn-link btn-sm me-3"
-                    >
-                      {" "}
-                      Eliminar
-                    </button>
-                  </div>
-                </td>
-              </tr>
-            ))
-          }
+          {sortedData().map((cliente, indice) => (
+            <tr key={indice}>
+              <th scope="row">{cliente.id}</th>
+              <td>{cliente.razonSocial}</td>
+              <td>{cliente.celular}</td>
+              <td>{cliente.mail}</td>
+              <td className="text-center">
+                <div>
+                  <Link to={`/cliente/${cliente.id}`} className="btn btn-link btn-sm me-3">
+                    Editar
+                  </Link>
+                  <button
+                    onClick={() => eliminar(cliente.id)}
+                    className="btn btn-link btn-sm me-3"
+                  >
+                    Eliminar
+                  </button>
+                </div>
+              </td>
+            </tr>
+          ))}
         </tbody>
       </table>
 
@@ -191,18 +175,19 @@ export default function ListadoCliente() {
           </Link>
         </div>
       </div>
+
       <div className="pagination d-md-flex justify-content-md-end">
-        {Array.from({ length: totalPages }, (_, i) => i).map((pageNumber) => (
-          <a
-            key={pageNumber}
-            href="#"
+        {Array.from({ length: totalPages }, (_, i) => (
+          <button
+            key={i}
+            className="btn btn-link"
             onClick={(e) => {
-              e.preventDefault(); // Previene el comportamiento predeterminado del enlace
-              handlePageChange(pageNumber);
+              e.preventDefault();
+              handlePageChange(i);
             }}
           >
-            | {pageNumber} |
-          </a>
+            {i + 1}
+          </button>
         ))}
       </div>
     </div>
